+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # development version
 
+- Add Snelder 2019 leuprorelin ([doi:10.1111/bcp.13891](https://doi.org/10.1111/bcp.13891)) -- men with prostate cancer.
+
 - Add Ousey 2026 plozasiran ([doi:10.1002/jcph.70190](https://doi.org/10.1002/jcph.70190)) -- adults with familial chylomicronemia syndrome.
 
 - Add Ravva 2010 varenicline exposure-response models ([doi:10.1038/clpt.2009.282](https://doi.org/10.1038/clpt.2009.282)) -- adult cigarette smokers in smoking-cessation trials.